gstreamer: Do not assume that the list index is the same as the device id.
This commit fixes a bug that was highlighted when devices had been added/removed or the backend was reloaded.
The AudioDevice used to use a static counter to allocate itself a device id that was propigated through the Phonon API.

Code in the Backend invalidly assumed that the index in the list was the same as this id.
In order to address this we add some helper methods to keep the AudioDevice.id field somewhat opaque.