pulseaudio: Rework the way in which PulseAudio support is included.
This commit changes the way in which pulseaudio support is indicated in the backend.
The API call of fullAudioDeviceEnumeration() has been removed and the backendinterface.h is now back
to how it was before (and thus no API change/breakage). The xine/gstreamer backends no longer
extend the PulseSupport class, instead accessing the PulseSupport object as a singleton.

This allows the PulseSupport to be accessed from various different parts of the code without
needing to change the backend interface class and break API.

Currently all the configuration is done via an exported API in libphononexperimental in
the BackendCapabilities. This interfaces to the config file at present but will soon be
adapted to make it also possibly a window onto PulseAudio if it is enabled.

I will probably need to find a better place to find to put the public interface than in
BackendCapabilities...