gstreamer: Fix a problem where the fact a reset was needed was lost.
If you create the path (createPath()) *before* setting the media source,
(setCurrentSource()) the gstreamer backend would forget the fact that a reset
was needed and ultimately end up in an error state.

This change simply does not wipe out the m_resetNeeded flag when
the source is set and leaves it as it is.

This fixes the test case application posted on
https://qa.mandriva.com/show_bug.cgi?id=56807