gstreamer: Fix a problem encountered when playing, stopping and playing again.
As reported by Harald Fernengel, a similar problem to that fixed in r1076454
also exists when you play->stop->play a media object.

This should fix the issue by marking a reset needed whenever we reach the stopped state.
Thanks to Harald for finding the problem and pointing the way to the fix.