pulsesupport: Make the backends explicitly state that they want PA support.
Now that we ask if pulse isactive and bypass asking the backends
for device lists etc if so, we cannot just assume that PA is supported
in the backend itself. So ensure that we enable pulseaudio support
when initialising the backends.

This was not needed before because we used to always go via the backend
for device information, even if it was just proxying on to the
PulseSupport class. This way it's more efficient and less
work to integrate into a backend.