gstreamer: Rejig the PulseAudio detection and handling a little.
Attempt to make the PulseAudio initialisation in GStreamer a bit more robust.
Validate the initial sink usage if it comes from either env var or settings
against the known status of PA (is it running or not) and only ever try a
'pulsesink' if it is running, with 'auto' favouring 'pulsesink' when
PA is running.

When trying to open the device fails when PA is running we change back
to 'auto' and try again. This handles the case when the GStreamer
'pulsesink' is not available (which is hard as it's a -good plugin
these days).

CCBUG: 59052