gstreamer: Move some code inside pointer protection.
This can only happen if asserts are disabled, so this is somewhat
unlikely, but it's better to be safe.

This is unlikely to be related to the bug #59052, but in the event
that PA is running but both 'pulsesink' and 'fakesink' cannot be
created AND asserts are disabled, then this could cause some issues.

CCBUG: 59052