GST: Second part of dvdmenu support - mouse over
Once the Gstreamer navigation interface processed a mouse event that
indicates that a mouse is over a clickable area it issues a message on
the bus. Same goes for when the mouse leaves a clickable area.
This GST bus message is converted into a MediaNodeEvent, which is then
processed by the videowidget to change the mouse cursor accordingly.

(Please note that the actual mouse button press/release handling is
already implemented from an earlier commit)