Map gst equalizer name to the one used in pxine
This allows API consumers such as Amarok to actually maintain a working
equalizer with phonon gstreamer, even though gstreamer uses a different
name for the EQ than phonon xine. This ought to be fixed in spitfire as
we are targetting an enum based identification system there. So, ideally
application developers can then check the available effects for
Effect::Equalizer and request that from the backend. Much more scalable
than the current name-based approach.

What I'd like to see meanwhile is amarok check for both KEqualizer and
equalizer-10bands in the EngineController. The manual mapping is very
ugly and should go away ASAP.

CCMAIL: amarok-devel@kde.org