Simplify and reform device management a bit
Introduce DeviceInfo instead of separate (and identical) classes for
video capture and audio output, merge audio and video device lists
inside the device manager.

Simplify the way the device manager is used by the backend and audio
output. The code should be easier to understand.

All devices other than the default will be marked as advanced.