fix assertation fails due to gstreamer reading from deleted readers
instead of deleting a read in setsource it is now deleted on setupsource
which is a callback from gstreamer when the new source is being actually
set up. in particular we may still be playing on stream1 when setsource is
called with stream2 (e.g. set while abouttofinish was in progress), so
previously stream1 would have been deleted while gstreamer was still
reading, now it is not.

also start a newly created reader right away as gstreamer may do probing
calls while we are still working in setupsource, so our reader needs to be
ready to handle those. starting asap is fine here as the setup is only done
once we start loading the source anyway (targeting playing/paused, so
technically statewise we are somewhere between stopped and playing).

BUG: 307469