assert when we have no tee in an A/V medianode
theory right now is that gstreamer is kaput and fails to factorize the
tee, outside MN nothing accesses the tee so the only way it could become
0x0 (assuming correct construction) is destruction. however since objects
are destructed through the main thread threading problems ought not be the
issue here.

At the very least we'll have an explicit assert so the casual code reader
knows what is going wrong.

CCBUG: 294006